Readme updated, DB init script changed, like button query definition in App.js. Some changes in like displaying.

+++ db/db-init.sql
@@ -13,7 +13,7 @@ CREATE TABLE posts(
     title varchar,
     body varchar,
     urllink varchar,
-    likes numeric(10)
+    likes SMALLINT
 );
 -- TODO
 INSERT INTO posts(title, body, urllink, likes) VALUES ('aaaaa', 'much aaaa', 'some.other.url', 0);
